+++ src/test/java/de/samply/core/TeilerCoreTest.java
@@ -1,9 +1,10 @@
-package de.samply.ldm;
+package de.samply.core;
 
 import de.samply.EnvironmentTestUtils;
 import de.samply.converter.ConverterManager;
 import de.samply.converter.Format;
 import de.samply.csv.ContainersToCsvConverter;
+import de.samply.excel.ContainersToExcelConverter;
 import de.samply.fhir.BundleToContainersConverter;
 import de.samply.query.QueryManager;
 import de.samply.template.ConverterTemplateManager;
@@ -16,14 +17,14 @@ import org.junit.jupiter.api.Test;
 import reactor.core.publisher.Flux;
 
 @Disabled
-class LdmClientTest {
+class TeilerCoreTest {
 
   private final String converterXmlApplicationContextPath = "./converter/converter.xml";
   private final String templateDirectory = "./templates";
   private final String writeDirectory = "./output";
   private final String sourceId = "blaze-store";
   private final String converterTemplateId = "test-template1";
-  private LdmClient ldmClient;
+  private TeilerCore teilerCore;
 
   @BeforeEach
   void setUp() {
@@ -33,19 +34,21 @@ class LdmClientTest {
     ContainersToCsvConverter containersToCsvConverter = new ContainersToCsvConverter(
         converterTemplateUtils,
         writeDirectory);
+    ContainersToExcelConverter containersToExcelConverter = new ContainersToExcelConverter(30000000,
+        converterTemplateUtils, writeDirectory);
     BundleToContainersConverter bundleToContainersConverter = new BundleToContainersConverter();
     ConverterManager converterManager = new ConverterManager(bundleToContainersConverter,
-        containersToCsvConverter, converterXmlApplicationContextPath);
+        containersToCsvConverter, containersToExcelConverter, converterXmlApplicationContextPath);
     ConverterTemplateManager converterTemplateManager = new ConverterTemplateManager(
         templateDirectory);
     QueryManager queryManager = new QueryManager();
 
-    ldmClient = new LdmClient(converterManager, converterTemplateManager, queryManager);
+    teilerCore = new TeilerCore(converterManager, converterTemplateManager, queryManager);
   }
 
   @Test
-  void retrieveByQueryId() throws LdmClientException {
-    Flux<Path> resultFlux = ldmClient.retrieveByQuery(sourceId, "Patient", Format.FHIR_QUERY,
+  void retrieveByQueryId() throws TeilerCoreException {
+    Flux<Path> resultFlux = teilerCore.retrieveByQuery(sourceId, "Patient", Format.FHIR_QUERY,
         Format.CSV, converterTemplateId);
     resultFlux.blockLast();
   }